% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreelCatch-package.R
\docType{package}
\name{CreelCatch-package}
\alias{CreelCatch}
\alias{CreelCatch-package}
\title{CreelCatch: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matthew Robertson \email{mdrobertson24@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
