% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prep.R
\name{model_prep}
\alias{model_prep}
\title{Prepare data and parameters for model run}
\usage{
model_prep(
  n_covars = 0,
  E_intercept = 1,
  C_intercept = 0,
  catchability = 0,
  catch = 0,
  effort = 0,
  random_effect = 0,
  first_covar = 0,
  second_covar = 0,
  third_covar = 0,
  gl_switch = 0,
  gl_area = 0,
  iGL = 0
)
}
\arguments{
\item{n_covars}{Number of covariates to explain fishing effort (integer)}

\item{E_intercept}{Binary to determine whether the effort intercept should be treated as a random effect. 0 indicates estimate as fixed effect, 1 indicates estimate as random effect. (binary)}

\item{C_intercept}{Scalar to determine whether the catch intercept should be estimated and if so if it should be estimated as a random effect. 0 indicates do not estimate, 1 indicates estimate as fixed effect, 2 indicates estimate as random effect. (scalar)}

\item{catchability}{Binary to determine whether the catchability coefficient should be treated as a random effect. 0 indicates estimate as fixed effect, 1 indicates estimate as random effect. (binary)}

\item{catch}{Input for desired catch data. No transformations are applied, so log(catch) should be input if that is the desired relationship to test. (numeric)}

\item{effort}{Input for desired effort data. No transformations are applied, so log(effort) should be input if that is the desired relationship to test. (numeric)}

\item{random_effect}{Input for random effect data. Should be the same length as catch and effort data. (integer)}

\item{first_covar}{Input for first covariate data. (numeric)}

\item{second_covar}{Input for second covariate data. (numeric)}

\item{third_covar}{Input for third covariate data. Note that current package functionality only allows this covariate to effect the model if not estimating separate Great Lakes waterbody area slopes (numeric)}

\item{gl_switch}{Binary to determine whether there should be different slopes between Great Lakes and non-Great Lakes waterbody area influence. 0 indicates only one slope, 1 indicates separate slopes (binary)}

\item{gl_area}{Input for waterbody area data if estimating different slopes between Great Lakes and non-Great Lakes waterbodies. (numeric)}

\item{iGL}{Binary vector indicating whether a waterbody is from the Great Lakes (1) or not (0). (binary vector)}
}
\value{
list
}
\description{
Prepare data and parameters for model run
}
\examples{
model_prep(n_covars = 0, E_intercept = 1, C_intercept = 0, catchability = 0,
catch = log(useable_dat$cpd), effort = log(useable_dat$epd), random_effect = useable_dat$istate,
first_covar = useable_dat$all_dev, second_covar = useable_dat$forest_wetland,
third_covar = useable_dat$Income, gl_area = useable_dat$Area, gl_switch = 0, iGL = useable_dat$GL)
}
