% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covars.R
\name{plot_covars}
\alias{plot_covars}
\title{Plot the covariate estimates}
\usage{
plot_covars(
  rep,
  sdrep,
  output,
  n_covar = 1,
  gl_switch = 0,
  covar_name = c("Covar1")
)
}
\arguments{
\item{rep}{input the report object from the model}

\item{sdrep}{input the sdreport object from the model}

\item{output}{input the output object from the model_prep function}

\item{n_covar}{input the number of covariates used in the model}

\item{gl_switch}{input whether the model had separate slopes estimated for Great Lakes (1) or not (0)}

\item{covar_name}{input a vector of the names of covariates}
}
\value{
plot
}
\description{
Plot the covariate estimates
}
\examples{
plot_covars(rep, sdrep, n_covar=2, gl_switch=0, covar_name=c("Area","Distance"))
}
