% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpue.R
\name{plot_cpue}
\alias{plot_cpue}
\title{Plot the relationship between catch and effort}
\usage{
plot_cpue(rep, sdrep, tmb.data, xlim = c(-1, 11), ylim = c(-1, 11))
}
\arguments{
\item{rep}{input the report object from the model}

\item{sdrep}{input the sdreport object from the model}

\item{tmb.data}{input the tmb.data object that was used to create the model (e.g. output$tmb.data)}

\item{xlim}{decide what the xlim of the figure will be (min, max). (optional)}

\item{ylim}{decide what the ylim of the figure will be (min, max). (optional)}
}
\value{
plot
}
\description{
Plot the relationship between catch and effort
}
\examples{
plot_cpue(rep, sdrep, tmb.data, xlim=c(-1,5), ylim=c(-1,5))
}
