% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_re.R
\name{plot_re}
\alias{plot_re}
\title{Plot the estimates and 95\% confidence intervals of random effects}
\usage{
plot_re(
  output,
  sdrep,
  names = c("Michigan", "Florida", "Connecticut", "Kansas", "South Dakota"),
  re_name = "State"
)
}
\arguments{
\item{output}{input the output object from the model_prep function}

\item{sdrep}{input the sdrep object}

\item{names}{list names of each level of the random effects (defaults to states in the example data)}

\item{re_name}{name of the random effect (defaults to "State")}
}
\value{
plot
}
\description{
Plot the estimates and 95\% confidence intervals of random effects
}
\examples{
plot_re(output, sdrep, names=c("Michigan","Florida","Connecticut","Kansas","South Dakota"), re_name="State")
}
